Adecuación de Grabado de Tarifas.

+++ src/main/java/pe/com/isesystem/siscopetarifario/model/TarifarioFleteId.java
@@ -21,7 +21,7 @@ public class TarifarioFleteId implements Serializable {
     private String codUbigeoDestino;
 
     @Column(name = "id_dia", nullable = false)
-    private Integer idDia;
+    private Long idDia;
 
     @Override
     public boolean equals(Object o) {
